Eliminating ncbitaxon:root and instead bridging to true organism subclasses
**please read closely and read the related issue before commenting**

This PR removed the statement that COB:organism = NCBITaxon:root

Rationale: root includes non-organisms such as samples

See https://github.com/obophenotype/ncbitaxon/issues/10

This PR replaces that link with two subClassOf axioms

 * COB:organism
    * NCBITaxon:Viruses
    * NCBITaxon:cellular organisms

Note that if this were merged, then the union of NCBITaxon and COB would have a lattice at the top. However, it would be trivial to add NCBITaxon:1 to an anti-slim and filter it out, resulting in a tree with COB:organism as the single MRCA of all organism classes

Merging this PR has some advantages such as obviating the need to rename NCBITaxon:1, since we instead simply inject our own true parent. It leaves the structure of labeling of NCBITaxon intact

**Important** please refrain from discussing the definition of organism here, specifically whether it includes viruses. For COB purposes, this is a closed issue. See: #6